Move Function: combineMeetingRoomsAndFloors()
